Update a LOT (again)
- Update a TON of coding
- Still have a lot of work to do > <"

+++ app/controllers/admin/Criteria.java
@@ -12,36 +12,61 @@ import views.html.*;
 public class Criteria extends Controller {
 
 	public static Result index() {
-		return ok(views.html.criteria_admin.render(Criterion.find.all(), form(Criterion.class)));
+		User user = User.find.where().eq("username", request().username()).findUnique();
+		if (user.isAdmin) {
+			return ok(views.html.criteria_admin.render(user, Criterion.find.all()));
+		}
+		else {
+			return redirect("/");
+		}
 	}
-	
+
 	public static Result criterion(Long id) {
+		User user = User.find.where().eq("username", request().username()).findUnique();
 		Criterion criterion = Criterion.find.where().eq("id", id).findUnique();
 		Form<Criterion> editForm = form(Criterion.class);
-		if (criterion != null) {
-			editForm = editForm.fill(criterion);
+		if (user.isAdmin) {
+			return ok(views.html.criterion_admin.render(user, criterion, criterion.ballotThisCriterion()));
 		}
-		return ok(views.html.criterion_admin.render(criterion, criterion.ballotThisCriterion(), editForm));
+		else {
+			return redirect("/");
+		}
 	}
-	
+
 	public static Result create() {
+		User user = User.find.where().eq("username", request().username()).findUnique();
 		Form<Criterion> createForm = form(Criterion.class).bindFromRequest();
 		createForm.get().save();
-		return ok(views.html.criteria_admin.render(Criterion.find.all(), form(Criterion.class)));
+		if (user.isAdmin) {
+			return ok(views.html.criteria_admin.render(user, Criterion.find.all()));
+		}
+		else {
+			return redirect("/");
+		}
 	}
-	
+
 	public static Result update(Long id) {
+		User user = User.find.where().eq("username", request().username()).findUnique();
 		Criterion criterion = Criterion.find.ref(id);
 		Form<Criterion> updateForm = form(Criterion.class).bindFromRequest();
 		Criterion updated = updateForm.get();
 		updated.update(criterion.id);
-		return ok(views.html.criterion_admin.render(updated, updated.ballotThisCriterion(), updateForm));
+		if (user.isAdmin) {
+			return ok(views.html.criterion_admin.render(user, updated, updated.ballotThisCriterion()));
+		}
+		else {
+			return redirect("/");
+		}
 	}
-	
+
 	public static Result delete(Long id) {
+		User user = User.find.where().eq("username", request().username()).findUnique();
+		if (!user.isAdmin) {
+			return redirect("/");
+		}
 		Criterion criterion = Criterion.find.ref(id);
 		criterion.delete();
 		return redirect("/admin/criterion");
 	}
-	
+
 }
